pyproject.toml
[project]
name = "dpf-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.12"
dependencies = [
    "cryptography==46.0.5",
    "django==6.0.2",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.16.1",
    "pymysql==1.1.2",
    "python-dotenv==1.2.1",
]
